FEAT Initialize a project and build a static website

+++ hub.py
@@ -14,8 +14,13 @@ app = typer.Typer()
 
 
 @app.command()
-def init(path:str):
+def init(path: str):
     typer.echo("Creating a new HUB site")
+    # Check if path exists
+    if os.path.exists(path):
+        typer.echo("Directory already exists")
+        return
+    os.mkdir(path)
     # Create a 'content' directory
     os.mkdir(f"{path}/content")
     with open(f"{path}/content/index.md", "w") as f:
@@ -24,6 +29,16 @@ def init(path:str):
         f.close()
     # Create a 'build' directory
     os.mkdir(f"{path}/build")
+    # Create a 'css' directory
+    os.mkdir(f"{path}/build/css")
+    # Create a 'default.css' file
+    with open(f"{path}/build/css/default.css", "w") as f:
+        f.write("body {\n")
+        f.write("    font-family: sans-serif;\n")
+        f.write("    background-color: black;\n")
+        f.write("    color: white;\n")
+        f.write("}\n")
+        f.close()
     # Create a 'config.yaml' file
     with open(f"{path}/config.yaml", "w") as f:
         f.write("title: My new HUB Site\n")
@@ -35,25 +50,39 @@ def init(path:str):
         f.write("index: index.md\n")
         f.close()
 
+
 @app.command()
-def build():
+def build(project_path: str):
+    """
+    Create static html files from the markdown files in the 'content' directory.
+    """
     typer.echo("Building site")
-    # Turn markdown files in 'content' directory into HTML
-    # Iterate over the files in the 'content' directory
-    for filename in os.listdir("content"):
+    for filename in os.listdir(f"{project_path}/content"):
         # Check if the file is a markdown file
         if filename.endswith(".md"):
             # Read the markdown file
-            with open("content/" + filename, "r") as f:
+            with open(f"{project_path}/content/" + filename, "r") as f:
                 content = f.read()
                 f.close()
             # Convert the markdown file to HTML
             html = markdown.markdown(content)
+            # Add css link based on the theme in the config file
+            theme = "default"
+            with open(f"{project_path}/config.yaml", "r") as f:
+                # Read theme parameter from YAML file
+                for line in f:
+                    if line.startswith("theme"):
+                        theme = line.split(":")[1].strip()
+                        break
+                f.close()
+            # Add the css link to the HTML
+            html = f'<link rel="stylesheet" href="css/{theme}.css">\n' + html
             # Write the HTML to a file in the 'build' directory
-            with open("build/" + filename.replace(".md", ".html"), "w") as f:
+            with open(
+                f"{project_path}/build/" + filename.replace(".md", ".html"), "w"
+            ) as f:
                 f.write(html)
                 f.close()
-    
 
 
 if __name__ == "__main__":
